i3/ipc: implement IPC listener to receive arbitrary events
