+++ src/x11/i3/ipc/qml.cpp
@@ -7,46 +7,49 @@
 #include "../../../core/model.hpp"
 #include "../../../core/qmlscreen.hpp"
 #include "connection.hpp"
+#include "controller.hpp"
 #include "workspace.hpp"
 
 namespace qs::i3::ipc {
 
 I3IpcQml::I3IpcQml() {
-	auto* instance = I3Ipc::instance();
+	auto* instance = I3IpcController::instance();
 
 	// clang-format off
 	QObject::connect(instance, &I3Ipc::rawEvent, this, &I3IpcQml::rawEvent);
 	QObject::connect(instance, &I3Ipc::connected, this, &I3IpcQml::connected);
-	QObject::connect(instance, &I3Ipc::focusedWorkspaceChanged, this, &I3IpcQml::focusedWorkspaceChanged);
-	QObject::connect(instance, &I3Ipc::focusedMonitorChanged, this, &I3IpcQml::focusedMonitorChanged);
+	QObject::connect(instance, &I3IpcController::focusedWorkspaceChanged, this, &I3IpcQml::focusedWorkspaceChanged);
+	QObject::connect(instance, &I3IpcController::focusedMonitorChanged, this, &I3IpcQml::focusedMonitorChanged);
 	// clang-format on
 }
 
-void I3IpcQml::dispatch(const QString& request) { I3Ipc::instance()->dispatch(request); }
-void I3IpcQml::refreshMonitors() { I3Ipc::instance()->refreshMonitors(); }
-void I3IpcQml::refreshWorkspaces() { I3Ipc::instance()->refreshWorkspaces(); }
-QString I3IpcQml::socketPath() { return I3Ipc::instance()->socketPath(); }
-ObjectModel<I3Monitor>* I3IpcQml::monitors() { return I3Ipc::instance()->monitors(); }
-ObjectModel<I3Workspace>* I3IpcQml::workspaces() { return I3Ipc::instance()->workspaces(); }
+void I3IpcQml::dispatch(const QString& request) { I3IpcController::instance()->dispatch(request); }
+void I3IpcQml::refreshMonitors() { I3IpcController::instance()->refreshMonitors(); }
+void I3IpcQml::refreshWorkspaces() { I3IpcController::instance()->refreshWorkspaces(); }
+QString I3IpcQml::socketPath() { return I3IpcController::instance()->socketPath(); }
+ObjectModel<I3Monitor>* I3IpcQml::monitors() { return I3IpcController::instance()->monitors(); }
+ObjectModel<I3Workspace>* I3IpcQml::workspaces() {
+	return I3IpcController::instance()->workspaces();
+}
 
 QBindable<I3Workspace*> I3IpcQml::bindableFocusedWorkspace() {
-	return I3Ipc::instance()->bindableFocusedWorkspace();
+	return I3IpcController::instance()->bindableFocusedWorkspace();
 }
 
 QBindable<I3Monitor*> I3IpcQml::bindableFocusedMonitor() {
-	return I3Ipc::instance()->bindableFocusedMonitor();
+	return I3IpcController::instance()->bindableFocusedMonitor();
 }
 
 I3Workspace* I3IpcQml::findWorkspaceByName(const QString& name) {
-	return I3Ipc::instance()->findWorkspaceByName(name);
+	return I3IpcController::instance()->findWorkspaceByName(name);
 }
 
 I3Monitor* I3IpcQml::findMonitorByName(const QString& name) {
-	return I3Ipc::instance()->findMonitorByName(name);
+	return I3IpcController::instance()->findMonitorByName(name);
 }
 
 I3Monitor* I3IpcQml::monitorFor(QuickshellScreenInfo* screen) {
-	return I3Ipc::instance()->monitorFor(screen);
+	return I3IpcController::instance()->monitorFor(screen);
 }
 
 } // namespace qs::i3::ipc
