+++ src/x11/i3/module.md
@@ -2,8 +2,10 @@ name = "Quickshell.I3"
 description = "I3 specific Quickshell types"
 headers = [
 	"ipc/connection.hpp",
+	"ipc/controller.hpp",
 	"ipc/qml.hpp",
 	"ipc/workspace.hpp",
 	"ipc/monitor.hpp",
+	"ipc/listener.hpp",
 ]
 -----
